fix CateService & ProdService

+++ src/main/java/com/spring/ecommerce/controller/ProductController.java
@@ -4,6 +4,8 @@ import com.spring.ecommerce.persistence.model.Product;
 import com.spring.ecommerce.service.ProductService;
 import com.spring.ecommerce.util.RestResponse;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -33,13 +35,40 @@ public class ProductController {
     /**Update product*/
     @PutMapping("/{productId}")
     public RestResponse updateProduct(@PathVariable("productId") Long productId, @RequestBody Product product) {
+//        try{
+//
+//        }
         return RestResponse.builder(productService.update(productId, product)).message("Success").build();
     }
+
+//    @DeleteMapping("/{productId}")
+//    public ResponseEntity <String> deleteProduct(@PathVariable Long productId){
+//        try {
+//            if (productService.findById(productId).isPresent()){
+//                productService.delete(productId);
+//                return ResponseEntity.status(HttpStatus.OK).body("Deleted product by id : " + productId);
+//            }else {
+//                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Product with id : " + productId + " not found");
+//            }
+//        }catch (Exception e){
+//            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
+//        }
+//
+//    }
+
 
     /**Delete product by id*/
     @DeleteMapping("/{productId}")
     public RestResponse deleteProductById(@PathVariable("productId") Long productId) {
-        productService.delete(productId);
-        return RestResponse.builder("Deleted product by id : " +productId).message("Success").build();
+        try {
+            if (productService.findById(productId).isPresent()) {
+                productService.delete(productId);
+                return RestResponse.builder("Deleted product by id : " + productId).message("Success").build();
+            }else {
+                return RestResponse.builder("Product with id " + productId + " not found !").message("NOT FOUND").status(404).build();
+            }
+        }catch (Exception e){
+            return RestResponse.builder(null).message(e.getMessage()).build();
+        }
     }
 }
